LPS-130175 Move ControlsContext component
