+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/Layout.js
@@ -27,9 +27,13 @@ import {LAYOUT_DATA_ITEM_TYPES} from '../config/constants/layoutDataItemTypes';
 import {LAYOUT_TYPES} from '../config/constants/layoutTypes';
 import {config} from '../config/index';
 import {useSetCollectionActiveItemContext} from '../contexts/CollectionActiveItemContext';
+import {
+	useActivationOrigin,
+	useIsActive,
+	useSelectItem,
+} from '../contexts/ControlsContext';
 import {useSelector} from '../store/index';
 import {deepEqual} from '../utils/checkDeepEqual';
-import {useActivationOrigin, useIsActive, useSelectItem} from './Controls';
 import FragmentWithControls from './layout-data-items/FragmentWithControls';
 import {
 	CollectionItemWithControls,
