+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/LayoutViewport.js
@@ -18,9 +18,9 @@ import React, {useEffect, useRef, useState} from 'react';
 import debounceRAF from '../../core/debounceRAF';
 import {VIEWPORT_SIZES} from '../config/constants/viewportSizes';
 import {config} from '../config/index';
+import {useSelectItem} from '../contexts/ControlsContext';
 import {GlobalContextFrame} from '../contexts/GlobalContext';
 import {useSelector} from '../store/index';
-import {useSelectItem} from './Controls';
 import DisabledArea from './DisabledArea';
 import Layout from './Layout';
 import MasterLayout from './MasterLayout';
