+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/contexts/CollectionActiveItemContext.js
@@ -14,15 +14,15 @@
 
 import React, {useContext, useState} from 'react';
 
-import {
-	useActiveItemId,
-	useActiveItemType,
-	useIsActive,
-} from '../components/Controls';
 import {ITEM_TYPES} from '../config/constants/itemTypes';
 import {LAYOUT_DATA_ITEM_TYPES} from '../config/constants/layoutDataItemTypes';
 import {useSelectorCallback} from '../store';
 import {CollectionItemContext, INITIAL_STATE} from './CollectionItemContext';
+import {
+	useActiveItemId,
+	useActiveItemType,
+	useIsActive,
+} from './ControlsContext';
 
 const CollectionActiveItemDispatchContext = React.createContext(() => {});
 const CollectionActiveItemStateContext = React.createContext(INITIAL_STATE);
