+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/comments/components/CommentsSidebar.js
@@ -14,9 +14,12 @@
 
 import React from 'react';
 
-import {useActiveItemId, useSelectItem} from '../../../app/components/Controls';
 import {HIGHLIGHTED_COMMENT_ID_KEY} from '../../../app/config/constants/highlightedCommentIdKey';
 import {LAYOUT_DATA_ITEM_TYPES} from '../../../app/config/constants/layoutDataItemTypes';
+import {
+	useActiveItemId,
+	useSelectItem,
+} from '../../../app/contexts/ControlsContext';
 import {useSelector} from '../../../app/store/index';
 import FragmentComments from './FragmentComments';
 import FragmentEntryLinksWithComments from './FragmentEntryLinksWithComments';
