+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/comments/components/FragmentComments.js
@@ -18,7 +18,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 import editFragmentEntryComment from '../../../app/actions/editFragmentEntryLinkComment';
-import {useSelectItem} from '../../../app/components/Controls';
+import {useSelectItem} from '../../../app/contexts/ControlsContext';
 import {useDispatch, useSelector} from '../../../app/store/index';
 import SidebarPanelContent from '../../../common/components/SidebarPanelContent';
 import SidebarPanelHeader from '../../../common/components/SidebarPanelHeader';
