+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/comments/components/FragmentEntryLinksWithComments.js
@@ -14,8 +14,11 @@
 
 import React from 'react';
 
-import {useHoverItem, useSelectItem} from '../../../app/components/Controls';
 import {LAYOUT_DATA_ITEM_TYPES} from '../../../app/config/constants/layoutDataItemTypes';
+import {
+	useHoverItem,
+	useSelectItem,
+} from '../../../app/contexts/ControlsContext';
 import {useSelector} from '../../../app/store/index';
 import SidebarPanelContent from '../../../common/components/SidebarPanelContent';
 import SidebarPanelHeader from '../../../common/components/SidebarPanelHeader';
