+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/contents/components/PageContent.js
@@ -22,9 +22,12 @@ import {openModal} from 'frontend-js-web';
 import PropTypes from 'prop-types';
 import React, {useEffect, useState} from 'react';
 
-import {useHoverItem, useHoveredItemId} from '../../../app/components/Controls';
 import {EDITABLE_FRAGMENT_ENTRY_PROCESSOR} from '../../../app/config/constants/editableFragmentEntryProcessor';
 import {ITEM_TYPES} from '../../../app/config/constants/itemTypes';
+import {
+	useHoverItem,
+	useHoveredItemId,
+} from '../../../app/contexts/ControlsContext';
 import {useSelector} from '../../../app/store/index';
 
 export default function PageContent(props) {
