+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/components/Topper.test.js
@@ -18,11 +18,11 @@ import React from 'react';
 import {DndProvider} from 'react-dnd';
 import {HTML5Backend} from 'react-dnd-html5-backend';
 
-import {ControlsProvider} from '../../../../src/main/resources/META-INF/resources/page_editor/app/components/Controls';
 import Topper from '../../../../src/main/resources/META-INF/resources/page_editor/app/components/Topper';
 import Row from '../../../../src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/Row';
 import {LAYOUT_DATA_ITEM_TYPES} from '../../../../src/main/resources/META-INF/resources/page_editor/app/config/constants/layoutDataItemTypes';
 import {VIEWPORT_SIZES} from '../../../../src/main/resources/META-INF/resources/page_editor/app/config/constants/viewportSizes';
+import {ControlsProvider} from '../../../../src/main/resources/META-INF/resources/page_editor/app/contexts/ControlsContext';
 import {StoreAPIContextProvider} from '../../../../src/main/resources/META-INF/resources/page_editor/app/store';
 
 jest.mock(
