+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/plugins/comments/components/FragmentEntryLinksWithComments.test.js
@@ -16,11 +16,11 @@ import '@testing-library/jest-dom/extend-expect';
 import {cleanup, fireEvent, render, wait} from '@testing-library/react';
 import React from 'react';
 
+import {LAYOUT_DATA_ITEM_TYPES} from '../../../../../../src/main/resources/META-INF/resources/page_editor/app/config/constants/layoutDataItemTypes';
 import {
 	useHoverItem,
 	useSelectItem,
-} from '../../../../../../src/main/resources/META-INF/resources/page_editor/app/components/Controls';
-import {LAYOUT_DATA_ITEM_TYPES} from '../../../../../../src/main/resources/META-INF/resources/page_editor/app/config/constants/layoutDataItemTypes';
+} from '../../../../../../src/main/resources/META-INF/resources/page_editor/app/contexts/ControlsContext';
 import {StoreAPIContextProvider} from '../../../../../../src/main/resources/META-INF/resources/page_editor/app/store/index';
 import FragmentEntryLinksWithComments from '../../../../../../src/main/resources/META-INF/resources/page_editor/plugins/comments/components/FragmentEntryLinksWithComments';
 
@@ -30,7 +30,7 @@ jest.mock(
 );
 
 jest.mock(
-	'../../../../../../src/main/resources/META-INF/resources/page_editor/app/components/Controls',
+	'../../../../../../src/main/resources/META-INF/resources/page_editor/app/contexts/ControlsContext',
 	() => {
 		const hoverItem = jest.fn();
 		const selectItem = jest.fn();
